remove unnecessary render function
