+++ frontend/packages/ux-editor/src/components/config/editModal/EditDataModelBindings.tsx
@@ -1,10 +1,12 @@
 import { useSelector } from 'react-redux';
-import { renderSelectDataModelBinding } from '../../../utils/render';
+import { PropertyLabel } from '../../../utils/render';
 import type { IAppState } from '../../../types/global';
 import type { IGenericEditComponent } from '../componentConfig';
 import { getMinOccursFromDataModel, getXsdDataTypeFromDataModel } from '../../../utils/datamodel';
 import { ComponentTypes } from '../../index';
-import { textSelector } from '../../../selectors/textSelectors';
+import React from 'react';
+import { useText } from '../../../hooks';
+import { SelectDataModelComponent } from '../SelectDataModelComponent';
 
 export interface EditDataModelBindingsProps extends IGenericEditComponent {
   renderOptions?: {
@@ -20,8 +22,8 @@ export const EditDataModelBindings = ({
   handleComponentChange,
   renderOptions,
 }: EditDataModelBindingsProps) => {
-  const language = useSelector(textSelector);
   const dataModel = useSelector((state: IAppState) => state.appData.dataModel.model);
+  const t = useText();
 
   const handleDataModelChange = (selectedDataModelElement: string, key = 'simpleBinding') => {
     handleComponentChange({
@@ -38,10 +40,23 @@ export const EditDataModelBindings = ({
     });
   };
 
-  return renderSelectDataModelBinding({
-    dataModelBinding: component.dataModelBindings,
-    onDataModelChange: handleDataModelChange,
-    language,
-    ...renderOptions,
-  });
+  const { uniqueKey, key, label } = renderOptions || {};
+  return (
+     <div key={uniqueKey || ''}>
+      <PropertyLabel
+        htmlFor={`selectDataModelSelect-${label}`}
+        textKey={
+          label
+            ? `${t('ux_editor.modal_properties_data_model_helper')} ${t('general.for')} ${label}`
+            : t('ux_editor.modal_properties_data_model_helper')
+        }
+      />
+      <SelectDataModelComponent
+        inputId={`selectDataModelSelect-${label}`}
+        selectedElement={component.dataModelBindings[key || 'simpleBinding']}
+        onDataModelChange={(dataModelField: string) => handleDataModelChange(dataModelField, key)}
+        noOptionsMessage={t('general.no_options')}
+      />
+    </div>
+  )
 };
